+++ src/routes/(app)/+page.server.ts
@@ -1,13 +1,9 @@
 import type { PageServerLoad } from './$types';
 import EventModel from '$lib/db/events';
 import type { Event } from '$lib/types/event.d.ts';
+import { eventService } from '$lib/services/eventService';
 
 export const load: PageServerLoad = async () => {
-	const events = (await EventModel.find({})
-		.sort({ dateTime: 'asc' })
-		.select(['-_id', '-__v'])
-		.lean()
-		.limit(4)) as unknown as Event[];
 
-	return { events };
+	return new eventService().getEvents();
 };
